dexie: more changing params back to int

+++ src/augr-fm/src/dexie_voice.cpp
@@ -135,9 +135,9 @@ void DexieVoice::LoadPatch(const Dx7Patch &patch) {
                                                kMaxDelaySeconds * sample_rate);
 
     for (int i = 0; i < 6; ++i) {
-        ops_[i]->lfo_amp_depth_ = static_cast<float>(patch.lfo_amp_depth);
-        ops_[i]->lfo_pitch_depth_ = static_cast<float>(patch.lfo_pitch_depth);
-        ops_[i]->pitch_mod_sens_ = static_cast<float>(patch.pitch_mod_sens);
+        ops_[i]->lfo_amp_depth_ = patch.lfo_amp_depth;
+        ops_[i]->lfo_pitch_depth_ = patch.lfo_pitch_depth;
+        ops_[i]->pitch_mod_sens_ = patch.pitch_mod_sens;
         ops_[i]->lfo_delay_samples_total_ =
             static_cast<int>((patch.lfo_delay / 99.0f) * kMaxDelaySeconds *
                              Audio::sample_rate());
@@ -219,7 +219,7 @@ void DexieVoice::PushOperatorParams(int op_idx, const Dx7Op &op, int feedback,
     d->kbd_left_curve_ = op.kbd_left_curve;
     d->kbd_right_curve_ = op.kbd_right_curve;
 
-    d->kbd_rate_scaling_ = static_cast<float>(op.kbd_rate_scaling);
+    d->kbd_rate_scaling_ = op.kbd_rate_scaling;
 
     d->fixed_freq_ = op.fixed_freq;
     d->frequency_ = op.fixed_freq ? FixedFrequencyHz(op.coarse_raw, op.fine_raw,
